Fix control UI browser bundle imports
